requirements.txt
pip>=25.1
setuptools>=70
wheel

fastapi>=0.110,<0.120
uvicorn>=0.27,<0.31

pydantic>=2.7.4,<3.0

python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
supabase==2.0.2
requests==2.31.0